feat : 장바구니 CRUD 생성

+++ src/main/java/org/example/delivery/domain/cart/dto/request/CreateCartItemRequest.java
@@ -15,5 +15,5 @@ public class CreateCartItemRequest {
     private Long menuId;    //메뉴 아이디
 
     @NotNull @Min(1)
-    private Long quantity;  //수량
+    private int quantity;  //수량
 }
